+++ cmd/books/root.go
@@ -14,7 +14,7 @@ import (
 
 var (
 	jsonOutput bool
-	version    = "0.2.0"
+	version    = "0.3.0"
 )
 
 var rootCmd = &cobra.Command{